Refactor: extracting functions to separate file
